[CHORE] Changed express gateway name
