+++ secure-gateway/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ecommerce-gateway",
+  "name": "secure-gateway",
   "description": "Express Gateway Instance Bootstraped from Command Line",
   "repository": {},
   "license": "UNLICENSED",
